+++ youwol/backends/assets_gateway/all_icons_emojipedia.py
