+++ youwol/backends/assets_gateway/models.py
