+++ youwol/backends/assets_gateway/routers/cdn.py
