+++ youwol/backends/assets_gateway/utils.py
