+++ youwol/middlewares/dynamic_routing/loading_graph_rules.py
@@ -6,12 +6,12 @@ import aiohttp
 from starlette.requests import Request
 from starlette.responses import Response, JSONResponse
 from starlette.middleware.base import RequestResponseEndpoint
-import youwol.services.backs.cdn.root_paths as cdn
+import youwol.backends.cdn.root_paths as cdn
 from youwol.middlewares.models_dispatch import AbstractDispatch
 from youwol_utils.context import Context
-from youwol.services.backs.cdn.configurations import get_configuration
+from youwol.backends.cdn.configurations import get_configuration
 from youwol.environment.youwol_environment import yw_config
-from youwol.services.backs.cdn.models import LoadingGraphBody
+from youwol.backends.cdn.models import LoadingGraphBody
 from youwol_utils import PackagesNotFound
